feat: add carapace package (#55)
* feat: Add buildParams and packagePath options to GoBuildOptions interface

The commit adds new options to the `GoBuildOptions` interface in the `go` package. The `buildParams` option allows specifying build parameters such as running `go generate` before building and passing ldflags to the `go install` command. The `packagePath` option allows setting the package path to build. This change enhances the flexibility and customization of the `goInstall` function.

Signed-off-by: jaudiger <jeremy.audiger@ioterop.com>

* feat: Add ldflagsWrapper function for escaping ldflags in shell commands and prefer to use env var when possible

Signed-off-by: jaudiger <jeremy.audiger@ioterop.com>

* feat: Add carapace package and dependencies

This commit adds the carapace package.

Signed-off-by: jaudiger <jeremy.audiger@ioterop.com>

* chore: Update carapace-bin URL to use carapace-sh organization

Signed-off-by: jaudiger <jeremy.audiger@ioterop.com>

* chore: Update ldflags in project.bri for carapace package

Signed-off-by: jaudiger <jeremy.audiger@ioterop.com>

---------

Signed-off-by: jaudiger <jeremy.audiger@ioterop.com>